split poc
